+++ tests/test_compound.py
@@ -132,6 +132,7 @@ def test_convolve():
     gsobject_compare(single, psf)
     check_basic(single, "`convolution' of single Moffat")
     do_pickle(single)
+    do_shoot(single, myImg, "single Convolution")
 
     single = galsim.Convolve([psf])
     gsobject_compare(single, psf)
@@ -589,6 +590,7 @@ def test_add():
     gsobject_compare(single, gauss1)
     check_basic(single, "`sum' of 1 Gaussian")
     do_pickle(single)
+    do_shoot(single, myImg, "Single Sum")
 
     single = galsim.Add([gauss1])
     gsobject_compare(single, gauss1)