+++ CHANGELOG.md
@@ -22,6 +22,14 @@ API Changes
 - Changed `InterpolatedKImage` to take an ImageC rather than two ImageD
   instances. The old syntax will work, but it will raise a deprecation
   warning. (#799)
+- Dynamic PhaseScreenPSFs now require an explicit start time and time step.
+  Clock management of phase screens now handled implicitly. (#824)
+- The time_step for updating atmospheric screens and the time_step used for
+  integrating a PhaseScreenPSF over time are now independent (#824)
+- OpticalScreen now requires `diam` argument. (#824)
+- Some of the backend (but nonetheless public API) methods of PhaseScreen and
+  PhaseScreenList have changed.  See the docstrings of these classes for
+  the new API if you have been using these methods. (#824)
 
 
 Dependency Changes
@@ -57,6 +65,9 @@ Deprecated Features
   sense.  If you had been using it, you should instead just divide the
   returned image by gain, which will have the same effect and probably
   be clearer in your own code about what you meant. (#799)
+- Deprecated ability to create multiple PhaseScreenPSFs with single call
+  to makePSF, since it's now just as efficient to call makePSF multiple
+  times. (#824)
 
 
 New Features
@@ -109,6 +120,10 @@ New Features
   random wavelengths from an SED. (#822)
 - Added function assignPhotonAngles to add arrival directions (in the form of
   dx/dz and dy/dz slopes) to an existing photon array. (#823)
+- Added geometric optics approximation for photon-shooting through
+  PhaseScreenPSFs (includes atmospheric PSF and OpticalPSF).  This
+  approximation is now the default for non-optical PhaseScreenPSFs. (#824)
+- Added gradient method to LookupTable2D. (#824)
 - Added `surface_ops` option to `drawImage` function, which applies a list of
   surface operations to the photon array before accumulating on the image.
   (#827)